Make distinct client and server sockets
Previously, the client socket and server socket were actually the
same socket instance, and they just emitted to themselves. This
works in many cases, but fails if the client and server both need
to process events with the same name. In that case, if the client
emitted the event, both the client and server would process the
event, while the intended behavior is that only the server should
process the event.

This adds a bit of complexity, but essentially creates two sockets,
where events emitted by one socket get processed by the other.